refactor: login page split into smaller component
